Mention google oauth plugin as a requirement
Under the "create a pipeline phase", there is a section
"Add your service account credentials" that refers to a
credential "Kind" of "Google Service Account from metadata".

This option is only availalbe if the google oauth plugin has
been installed.

This commit adds it to the list of required jenkins plugins.