Clarify how to poll and monitor /version (#40)
* Clarify how to poll and monitor /version

Problem:
The instruction on how to poll and monitor changes during deployment are not clear. For example:

Step 6 of "Deploy the sample app to Kubernetes" instructs us to create an environmental variable called FRONTEND_SERVICE_IP. We do this in the original Google Cloud Shell terminal, as this is where we've been running all of the commands up to this point, and have not been told to do otherwise.

Step 8 then tells us to open a terminal and run a command that references the environmental variable from step 6. However, when we do so, the new terminal does not have access to the environmental variable, because it is a different session. Instead of seeing the version output, we get a curl error saying that the host `http` cannot be resolved.

Fix:
I modified the documentation to instruct the user to open the new terminal *before* setting the environmental variable. This ensures that it is available for the command in step 8. I also added some clarification on what "open a terminal" meant, as I was not initially aware that it was possible to open multiple Google Cloud Shell terminals, and instead tried to execute the command on a terminal running on my workstation.

* Combine 'Deploy the sample app to Kubernetes' steps 7 & 8