Fix readme indentation & formatting of code blocks
Problem:
The code blocks are not formatted consistently. Since the file source is not typically
viewed, this isn't a huge problem. However, at least tree of the code blocks in
"Phase 5: Deploy a development branch" are not rendered correctly due to the formatting
issues:

* The "You should see that..." text after step 1 is not indented, which causes the numbered bullet points to reset.

* Step 1 is showing the `shell` tag for the code block as part of the code, and the line breaks are not being rendered.

* Step 2 (which is displayed as a second step 1) is also showing the `shell` tag for the code block.

Fix:

I reformatted the code blocks so they are all consistent:

* All blocks are indented with 3 spaces
* All blocks have a blank line above and below them