Fix branch and remote names in phase 5.
Problem:
Steps within phase 5 are referring to the wrong branch names when instructing the reader to deploy to staging and production.

For example, the document refers to a `production` branch, which is actually `master` (see Jenkinsfile:30).

Additionally, the git push commands did not include the remote name, and were instead just instructing the reader to `git push master`. This will only work if you have configured an upstream branch, which if the reader had followed the tutorial word for word, would not have.

Fix:
I corrected all of the branch names and added the missing remote argument to the git push commands.