Fix & clarify phase 5 (#43)
Problem:
The code blocks are not formatted consistently. Since the file source is not typically viewed, this isn't a huge problem. However, at least tree of the code blocks in "Phase 5: Deploy a development branch" are not rendered correctly due to the formatting issues:

* Step 1 of the deployment section is showing the `shell` tag for the code block as part of the code, and the line breaks are not being rendered.

* Step 1 of the access section is also showing the `shell` tag for the code block.

Additionally, the instructions to view the new job and it's console output are terse and vague. It is assumed that the reader is familiar with Jenkins and knows how to access these things.

Fix:

* I reformatted the code blocks in phase 5 so they are consistent
* All blocks are indented with 3 spaces
* All blocks have a blank line above and below them
* The "Deploy" vs "Access" stages of the "Deploy a development branch" have been more clearly defined, with full instructions on how to access the Console Output of the Jenkins build.