Fix securing jenkins breaks ingress. (#45)
* Fix securing jenkins breaks ingress.

Problem:
Securing the Jenkins installation by enabling matrix based authorization prevents the user from accessing Jenkins via the Ingress load balancer.

Cause:
The load balancer is polling the Jenkins host web root, which returns a 403 when authorization is not allowing anonymous access. This causes the Ingress load balancer to believe the host is not available.

Fix:
I added the readinessProbe to the Jenkins Deployment as suggested by @ryane. This causes Ingress to poll /login instead of the host web root, which responds with a 200 OK for anonymous requests.

Fixes #21

* Adjusted timings