Rename staging environment to canary.
Requirement:
Issue #11 confirms that the environment named staging in the repo is actually using the canary pattern, and should be named as such.

Implementation:
Mostly just a find and replace on the word staging vs canary. However I did ensure that the first instance of the word canary is a link to Martin Fowler's Canary Release article, and removed the statement claiming that staging (canary) and production environments are isolated.

Resolves #11